requirements.txt
allure-pytest==2.8.24
allure-python-commons==2.8.24
apipkg==1.5
appdirs==1.4.4
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
delayed-assert==0.3.2
dill==0.3.3
distlib==0.3.1
enum34==1.1.10
execnet==1.7.1
filelock==3.0.12
greenlet==1.0a1
idna==2.10
iniconfig==1.1.1
packaging==20.8
pipenv==2020.11.15
playwright==0.170.0
pluggy==0.13.1
py==1.10.0
pyee==8.1.0
pyparsing==2.4.7
pytest==6.2.1
pytest-forked==1.3.0
pytest-reportportal==5.0.7
pytest-xdist==2.2.0
reportportal-client==5.0.7
requests==2.25.1
selenium==3.141.0
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.2.2
virtualenv-clone==0.5.4
